refactor: 8단계 전 리뷰사항 개선
